+++ dart/geobase/CHANGELOG.md
@@ -14,6 +14,9 @@
 🛠 Refactoring:
 * [Remove dependency on equatable package #248](https://github.com/navibyte/geospatial/issues/248)
 
+🛠 Maintenance:
+* [Update very_good_analysis version to 6.0.0 #249](https://github.com/navibyte/geospatial/issues/249)
+
 ## 1.3.0
 
 ✨ New (2024-07-26): The stable version 1.3.0 with centroid, polylabel, point-in-polygon and other cartesian 2D calculations enhanced - [read more](https://geospatial.navibyte.dev/v1/geobase/geometry-calculations/)!